Initial refactor for multiple providers
