+++ .goreleaser.yml
@@ -16,7 +16,7 @@ builds:
       - CGO_ENABLED=0
     # Adds some metadata to the build like version, commit reference and the date of the build
     ldflags:
-      - -s -w -X "github.com/gogatekeeper/gatekeeper/pkg/proxy.release={{.Version}}" -X "github.com/gogatekeeper/gatekeeper/pkg/proxy.gitsha={{ .FullCommit }}" -X "github.com/gogatekeeper/gatekeeper/pkg/proxy.compiled={{ .Timestamp }}"
+      - -s -w -X "github.com/gogatekeeper/gatekeeper/pkg/proxy/core.release={{.Version}}" -X "github.com/gogatekeeper/gatekeeper/pkg/proxy/core.gitsha={{ .FullCommit }}" -X "github.com/gogatekeeper/gatekeeper/pkg/proxy/core.compiled={{ .Timestamp }}"
     main: ./cmd/gatekeeper-keycloak.go
     binary: gatekeeper
 
@@ -34,7 +34,7 @@ builds:
       - CGO_ENABLED=0
     # Adds some metadata to the build like version, commit reference and the date of the build
     ldflags:
-      - -s -w -X "github.com/gogatekeeper/gatekeeper/pkg/proxy.release={{.Version}}" -X "github.com/gogatekeeper/gatekeeper/pkg/proxy.gitsha={{ .FullCommit }}" -X "github.com/gogatekeeper/gatekeeper/pkg/proxy.compiled={{ .Timestamp }}"
+      - -s -w -X "github.com/gogatekeeper/gatekeeper/pkg/proxy/core.release={{.Version}}" -X "github.com/gogatekeeper/gatekeeper/pkg/proxy/core.gitsha={{ .FullCommit }}" -X "github.com/gogatekeeper/gatekeeper/pkg/proxy/core.compiled={{ .Timestamp }}"
     main: ./cmd/gatekeeper-keycloak.go
     binary: gatekeeper
 
@@ -46,7 +46,7 @@ builds:
       - CGO_ENABLED=0
     # Adds some metadata to the build like version, commit reference and the date of the build
     ldflags:
-      - -s -w -X "github.com/gogatekeeper/gatekeeper/pkg/proxy.release={{.Version}}" -X "github.com/gogatekeeper/gatekeeper/pkg/proxy.gitsha={{ .FullCommit }}" -X "github.com/gogatekeeper/gatekeeper/pkg/proxy.compiled={{ .Timestamp }}"
+      - -s -w -X "github.com/gogatekeeper/gatekeeper/pkg/proxy/core.release={{.Version}}" -X "github.com/gogatekeeper/gatekeeper/pkg/proxy/core.gitsha={{ .FullCommit }}" -X "github.com/gogatekeeper/gatekeeper/pkg/proxy/core.compiled={{ .Timestamp }}"
     main: ./cmd/gatekeeper-keycloak.go
     binary: gatekeeper
