+++ pkg/keycloak/proxy/middleware.go
