+++ pkg/keycloak/proxy/oauth_proxy.go
@@ -6,7 +6,6 @@ import (
 	"net"
 	"net/http"
 	"net/url"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -14,10 +13,9 @@ import (
 	"github.com/Nerzal/gocloak/v12"
 	oidc3 "github.com/coreos/go-oidc/v3/oidc"
 	"github.com/gogatekeeper/gatekeeper/pkg/authorization"
-	"github.com/gogatekeeper/gatekeeper/pkg/config"
 	"github.com/gogatekeeper/gatekeeper/pkg/constant"
+	"github.com/gogatekeeper/gatekeeper/pkg/keycloak/config"
 	"github.com/gogatekeeper/gatekeeper/pkg/storage"
-	"github.com/prometheus/client_golang/prometheus"
 	"go.uber.org/zap"
 	"gopkg.in/square/go-jose.v2/jwt"
 )
@@ -197,60 +195,4 @@ type UserContext struct {
 	Claims map[string]interface{}
 	// permissions
 	Permissions authorization.Permissions
-}
-
-var (
-	release  = ""
-	gitsha   = "no gitsha provided"
-	compiled = "0"
-	version  = ""
-)
-
-var (
-	certificateRotationMetric = prometheus.NewCounter(
-		prometheus.CounterOpts{
-			Name: "proxy_certificate_rotation_total",
-			Help: "The total amount of times the certificate has been rotated",
-		},
-	)
-	oauthTokensMetric = prometheus.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "proxy_oauth_tokens_total",
-			Help: "A summary of the tokens issuesd, renewed or failed logins",
-		},
-		[]string{"action"},
-	)
-	oauthLatencyMetric = prometheus.NewSummaryVec(
-		prometheus.SummaryOpts{
-			Name: "proxy_oauth_request_latency",
-			Help: "A summary of the request latancy for requests against the openid provider, in seconds",
-		},
-		[]string{"action"},
-	)
-	latencyMetric = prometheus.NewSummary(
-		prometheus.SummaryOpts{
-			Name: "proxy_request_duration",
-			Help: "A summary of the http request latency for proxy requests, in seconds",
-		},
-	)
-	statusMetric = prometheus.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "proxy_request_status_total",
-			Help: "The HTTP requests partitioned by status code",
-		},
-		[]string{"code", "method"},
-	)
-)
-
-// GetVersion returns the proxy version
-func GetVersion() string {
-	if version == "" {
-		tm, err := strconv.ParseInt(compiled, 10, 64)
-		if err != nil {
-			return "unable to parse compiled time"
-		}
-		version = fmt.Sprintf("%s (git+sha: %s, built: %s)", release, gitsha, time.Unix(tm, 0).Format("02-01-2006"))
-	}
-
-	return version
 }
